[docs/design] membership 컴포넌트 WatchingContent 파일 변경 / 경로 변경 / gsap 위치 조정중
